backend-python-ml/requirements.txt
grpcio>=1.62.0
grpcio-tools>=1.62.0
protobuf>=5.26.1
# NumPy must be <2.0 for compatibility with PyTorch 2.1.2
numpy<2.0
# Use newer faster-whisper with compatible PyAV
av==12.3.0
faster-whisper==1.0.3
transformers==4.36.0
torch>=2.2.0
torchaudio>=2.2.0
edge-tts==7.2.3
pydub==0.25.1
pyloudnorm==0.1.1
pyyaml==6.0.1
nltk==3.8.1
googletrans==4.0.0rc1
sentencepiece==0.1.99
langdetect==1.0.9
uvicorn[standard]==0.25.0
python-multipart==0.0.6
psutil==5.9.6
asyncio==3.4.3
aiofiles==23.2.1
scipy==1.11.4

